requirements.txt
Flask==2.3.3
Flask-WTF==1.1.1
WTForms==3.0.1
requests==2.31.0
python-dotenv==1.0.0
pandas==2.0.3
matplotlib==3.7.2
seaborn==0.12.2
beautifulsoup4==4.12.2
lxml==4.9.3
gunicorn==21.2.0
